Use more config
